+++ burguer-queen/src/app/adminViews/admin/admin.component.css
@@ -1,76 +1,25 @@
-.hide{
-    display: none;
-}
-a{
-    color: aquamarine;
-    text-align: center;
-    font-size: 27px;
-    font-family: 'Inter', sans-serif;
-    margin:20px;
-}
-img{
-    width: 100px;
-    height: 100px;
-}
-#modal{
-    margin-top: 10px;
-    padding: 10px;
-    background-color: #FB727C;
-    border: none;
-    margin-right: 10px;
-  }
-  /* Structure */
 
-.example-container {
+  .icons {
     position: relative;
-  }
-  
-  .example-table-container {
-    position: relative;
-    min-height: 200px;
-    max-height: 400px;
-    overflow: auto;
-  }
-  
-  table {
-    width: 100%;
-    max-width: 1200px;
-  }
-  
-  .example-loading-shade {
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 56px;
-    right: 0;
-    background: rgba(0, 0, 0, 0.15);
-    z-index: 1;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-  
-  .example-rate-limit-reached {
-    max-width: 360px;
+    left: 46%;
+    width: 80px;
+    height: 80px;
+    font-size: 50px;
     text-align: center;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  .icon {
+    position: relative;
+    left: 45%;
+    width: 120px;
+    height: 80px;
+    font-size: 20px;
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    top: -30px;
   }
   
-  /* Column Widths */
-  .mat-column-number,
-  .mat-column-state {
-    max-width: 64px;
-  }
-  
-  .mat-column-created {
-    max-width: 124px;
-  }
-.iconEmployees{
-  font-size:20px;
-  width: 80px;
-  height:80px;
-}
-.iconProductList{
-font-size:20px;
-width: 80px;
-height:80px;
-}
