+++ burguer-queen/src/app/adminViews/admin/admin.component.html
@@ -1,81 +1,9 @@
-<mat-icon class="iconEmployees" routerLink="employees" aria-hidden="false" aria-label="Example home icon">account_circle
+<mat-icon class="icons" routerLink="employees" aria-hidden="false" aria-label="Example home icon">account_circle
 </mat-icon>
-<p class="iconEmployees"> Ir a Lista de Empleados
+<p class="icon"> Ir a Lista de Empleados
 </p>
-<mat-icon class="iconProductList" aria-label="Example home icon"> fastfood</mat-icon>
-<p class="iconProductList"> Estas en Listado de Productos 
-</p>
+<mat-icon class="icons" routerLink="totalproduct" aria-label="Example home icon"> fastfood</mat-icon>
+<p class="icon"> Estas en Listado de Productos 
+</p> 
 
-<router-outlet></router-outlet>
-<!-- <mat-icon routerLink="employees" class= "iconEmployee" aria-hidden="false" aria-label="Example home icon">account_circle</mat-icon><br>
-<router-outlet></router-outlet> -->
-<button id="modal" mat-button (click)="openDialog()">Añadir al carrito</button>
-
-<div class="container">
-  <div style="margin-top: 10px;"></div>
-</div>
-<mat-form-field appearance="standard">
-  <mat-label>Filter</mat-label>
-  <input matInput (keyup)="applyFilter($event)" placeholder="Ex. Mia" #input>
-</mat-form-field>
-
-<div class="mat-elevation-z8">
-  <table mat-table [dataSource]="dataSource" matSort>
-
-    <!-- ID Column -->
-    <ng-container matColumnDef="id">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header> ID </th>
-      <td mat-cell *matCellDef="let row"> {{row.id}} </td>
-    </ng-container>
-
-
-    <ng-container matColumnDef="name">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header> Nombre </th>
-      <td mat-cell *matCellDef="let row"> {{row.name}} </td>
-    </ng-container>
-
-    <ng-container matColumnDef="price">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header> Precio </th>
-      <td mat-cell *matCellDef="let row"> {{row.price | currency }} </td>
-    </ng-container>
-
-     <ng-container matColumnDef="type">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header> Tipo </th>
-      <td mat-cell *matCellDef="let row"> {{row.type}} </td>
-    </ng-container>
-
-     <ng-container matColumnDef="image">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header> Imagen </th>
-      <td mat-cell *matCellDef="let row"><img src="{{row.image}}" alt="producto"></td>
-    </ng-container>
-
-    <ng-container matColumnDef="action">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header> Acciones </th>
-      <td mat-cell *matCellDef="let row">
-       <button mat-icon-button (click)="editProduct(row)" color="primary" >
-        <mat-icon>edit</mat-icon>
-       </button>
-       <button mat-icon-button color="warn" 
-       (click)="onProductDelete(row.id)"
-       ><mat-icon>delete</mat-icon>
-       </button>
-      </td>
-    </ng-container>
-
-  
-     
-    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-    <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-
-    <!-- Row shown when there is no matching data. -->
-    <tr class="mat-row" *matNoDataRow>
-      <td class="mat-cell" colspan="4">No data matching the filter "{{input.value}}"</td>
-    </tr>
-  </table>
-
-  <mat-paginator [pageSizeOptions]="[5, 10, 25, 100]" aria-label="Select page of users"></mat-paginator>
-</div>
-<button routerLink="" (click)="logOut()" mat-stroked-button color="warn">cerrar Sesion</button>
-
-<router-outlet></router-outlet>
-  
+<router-outlet></router-outlet>
