+++ burguer-queen/src/app/adminViews/employees/employees.component.html
@@ -28,6 +28,7 @@
 
 <mat-paginator [pageSizeOptions]="[5, 10, 25, 100]" aria-label="Select page of users"></mat-paginator>
 </div>
+<button routerLink="" (click)="logOut()" mat-stroked-button color="warn">cerrar Sesion</button>
 <router-outlet></router-outlet>
 
  <!-- <div class="mat-elevation-z8">
