+++ burguer-queen/src/app/app-routing.module.ts
@@ -13,6 +13,7 @@ import { ProductsComponent } from 'src/app/waiterView/products/products.componen
 import { PendingStatusComponent } from './ChefViews/pending/pending-status.component';
 import { DeliveredStatusComponent } from './ChefViews/delivered-status/delivered-status.component';
 import { ReadyOrdersComponent } from './waiterView/ready-orders/ready-orders.component';
+import { TotalproductComponent } from './adminViews/totalproduct/totalproduct.component';
 
 const routes: Routes = [
   { path: '',redirectTo: '/home', pathMatch: 'full'},
@@ -20,6 +21,7 @@ const routes: Routes = [
   { path: 'admin', component: AdminComponent,
   children: [ 
     {path:'employees',component: EmployeesComponent},
+    {path:'totalproduct',component: TotalproductComponent},
   ]
 },
   { path: 'products', component: ProductsComponent,
